Add bounds checks for ramp floor/ceiling Otherwise, we could theoretically set the floor or ceiling greater than MAX_LEVEL and set the ceiling to be less than the floor. Either of these would bork the step size/nearest level calculations.
On the other hand, the user would have to pretty deliberately do
something like this unless maybe the floor and ceiling are really low...
but it still seems important to enforce this.

[OVERFLOW WARNING]
The main downside is that this is somewhat expensive to check, so you'll
probably have to trade a strobe or something.